KCC_Query_Assistant/requirements.txt
streamlit==1.31.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
duckduckgo-search==1.8.9
torch==2.2.1
transformers==4.37.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
requests==2.31.0
